backend/requirements.txt
bs4==0.0.1
celery==5.0.5
Django==3.1.7
django-cors-headers==3.7.0
django-crontab==0.7.1
django-extensions==3.1.1
django-rest-framework==0.1.0
django-rest-jwt==0.1.46
djangorestframework==3.12.2
Pillow==8.2.0
requests==2.25.1
